modifications and bug fixed

+++ API/routes/shoppingCartRoute.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const shoppingCartModel = require("../models/shoppingCartModel");
 const router = express.Router();
+const createError = require("http-errors");
 
 // get all cart details
 router.get("/", async (req, res) => {
@@ -41,9 +42,9 @@ router.post("/", async (req, res, next) => {
       userId: req.body.userId,
       itemId: req.body.itemId,
       itemName: req.body.itemName,
-      itembrand: req.body.itembrand,
-      itemprice: req.body.itemprice,
-      itemimgUrl: req.body.itemimgUrl,
+      itemBrand: req.body.itemBrand,
+      itemPrice: req.body.itemPrice,
+      itemImgUrl: req.body.itemImgUrl,
       itemCount: req.body.itemCount,
     });
 
@@ -121,7 +122,10 @@ router.delete("/:itemId", async (req, res, next) => {
 router.delete("/deletecart/:userId", async (req, res, next) => {
   try {
     let items = await shoppingCartModel.find({ userId: req.params.userId });
-   
+
+    if (items.length === 0) {
+      throw createError(400, "No Items In The Cart");
+    }
     let phoneId = "";
     for (i = 0; i < items.length; i++) {
       phoneId = await shoppingCartModel.findOneAndDelete({
@@ -131,7 +135,6 @@ router.delete("/deletecart/:userId", async (req, res, next) => {
     }
     res.send(phoneId);
   } catch (error) {
-    //console.log(error.message);
     if (error.name === "ValidationError") {
       next(createError(422, error.message));
       return;
@@ -139,6 +142,8 @@ router.delete("/deletecart/:userId", async (req, res, next) => {
     next(error);
   }
 });
+
+
 // ProductModel.findOneAndDelete({ brand: 'Nike' }, function (err) {
 //   if(err) console.log(err);
 //   console.log("Successful deletion");
